requirements.txt
certifi==2026.2.25
dnspython==2.8.0
numpy==2.4.4
pandas==3.0.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
tqdm==4.67.3
